+++ beam-migrate/ChangeLog.md
@@ -1,3 +1,16 @@
+
+# Unreleased
+
+## Added features
+
+* Added support for foreign key constraints:
+
+  * New `ForeignKeyAction` datatype representing possible actions when updating
+    or deleting a row with referencing foreign keys.
+  * The `IsSql92TableConstraintSyntax` typeclass now has an additional method,
+    `foreignKeyConstraintSyntax`, for constructing foreign key constraint syntax.
+  * Introduce `addTableForeignKey` for declaring new foreign key constraints.
+
 # 0.5.4.0
 
 ## Added features